AbstractBookActivityを廃止してUtilsクラスを作成
